% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3--aaa-setup.R
\name{as_R6}
\alias{as_R6}
\title{As R6}
\usage{
as_R6(x)
}
\arguments{
\item{x}{any object. If it inherits \code{'R6'}, then the class of \code{x} is set to \code{'R6'}}
}
\description{
Debug method that strips all gqlr classes and assigns the class as \code{'R6'}
}
\examples{
print(Int)
print(as_R6(Int))
}
