% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{check_if_gqlr_object}
\alias{check_if_gqlr_object}
\title{formal function to make sure all of the objects are gqlr objects}
\usage{
check_if_gqlr_object(x, kind = class(x))
}
\description{
formal function to make sure all of the objects are gqlr objects
}

