% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution-field.R
\name{GetFieldEntry}
\alias{GetFieldEntry}
\title{Get Field Entry}
\usage{
GetFieldEntry(objectType, object, fields)
}
\description{
Each item in the grouped field set can potentially create an entry in the result map. That entry in the result map is the result is the result of calling GetFieldEntry on the corresponding item in the grouped field set. GetFieldEntry can return null, which indicates that there should be no entry in the result map for this item. Note that this is distinct from returning an entry with a string key and a null value, which indicates that an entry in the result should be added for that key, and its value should be null.
}
\details{
GetFieldEntry assumes the existence of two functions that are not defined in this section of the spec. It is expected that the type system provides these methods:

ResolveFieldOnObject, which takes an object type, a field, and an object, and returns the result of resolving that field on the object.

GetFieldTypeFromObjectType, which takes an object type and a field, and returns that field’s type on the object type, or null if the field is not valid on the object type.'
}

