% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_add_to_schema}
\alias{gqlr_add_to_schema}
\title{Add item to schema}
\usage{
gqlr_add_to_schema(schemaObj, obj)
}
\arguments{
\item{schemaObj}{schema object to add to}

\item{obj}{parsed object to add to the schemaObj}
}
\description{
Add item to schema
}
\examples{
defs <- test_json("kitchen_schema")$definitions
cat(test_string("kitchen_schema"), "\\n")
schemaObj <- gqlr_init_schema() \%>\%
  gqlr_add_to_schema(gqlr_parse(defs[[1]])) \%>\%
  gqlr_add_to_schema(gqlr_parse(defs[[2]])) \%>\%
  gqlr_add_to_schema(gqlr_parse(defs[[3]])) \%>\%
  gqlr_add_to_schema(gqlr_parse(defs[[4]])) \%>\%
  gqlr_add_to_schema(gqlr_parse(defs[[5]])) \%>\%
  gqlr_add_to_schema(gqlr_parse(defs[[6]])) \%>\%
  gqlr_add_to_schema(gqlr_parse(defs[[7]])) \%>\%
  gqlr_validate_schema()
str(schemaObj, 3)
}

