% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.default}
\alias{gqlr_parse.default}
\title{default method to help with implementing new types}
\usage{
\method{gqlr_parse}{default}(obj, ...)
}
\description{
default method to help with implementing new types
}

