% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.Document}
\alias{gqlr_parse.Document}
\title{parse graphql Document}
\usage{
\method{gqlr_parse}{Document}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
Evaluating requests
}
\details{
To evaluate a request, the executor must have a parsed Document (as defined in the “Query Language” part of this spec) and a selected operation name to run if the document defines multiple operations.

The executor should find the Operation in the Document with the given operation name. If no such operation exists, the executor should throw an error. If the operation is found, then the result of evaluating the request should be the result of evaluating the operation according to the “Evaluating operations” section.
}
\examples{
gqlr_parse(test_json("simple-film-schema"))
}

