% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.EnumTypeDefinition}
\alias{gqlr_parse.EnumTypeDefinition}
\title{parse graphql EnumTypeDefinition}
\usage{
\method{gqlr_parse}{EnumTypeDefinition}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql EnumTypeDefinition
}
\examples{
enumObj <- test_json("kitchen_schema")$definitions[[5]]
gqlr_parse(enumObj)
}

