% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.EnumValueDefinition}
\alias{gqlr_parse.EnumValueDefinition}
\title{parse graphql EnumValueDefinition}
\usage{
\method{gqlr_parse}{EnumValueDefinition}(x)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql EnumValueDefinition
}
\examples{
enumValueObj <- test_json("kitchen_schema")$definitions[[5]]$values[[1]]
gqlr_parse(enumValueObj)
}

