% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.FieldDefinition}
\alias{gqlr_parse.FieldDefinition}
\title{parse graphql Definition FieldDefinition}
\usage{
\method{gqlr_parse}{FieldDefinition}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql Definition FieldDefinition
}
\examples{
objObj <- test_json("kitchen_schema")$definitions[[1]]
fieldObj <- objObj$fields[[1]]
gqlr_parse(fieldObj)
}

