% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.InputObjectTypeDefinition}
\alias{gqlr_parse.InputObjectTypeDefinition}
\title{parse graphql InputObjectTypeDefinition}
\usage{
\method{gqlr_parse}{InputObjectTypeDefinition}(x)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql InputObjectTypeDefinition
}
\examples{
inputObj <- test_json("kitchen_schema")$definitions[[6]]
gqlr_parse(inputObj)
}

