% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.InputValueDefinition}
\alias{gqlr_parse.InputValueDefinition}
\title{parse graphql Definition InputValueDefinition}
\usage{
\method{gqlr_parse}{InputValueDefinition}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql Definition InputValueDefinition
}
\examples{
inputValObj <- test_json("kitchen_schema")$definitions[[6]]$fields[[1]]
gqlr_parse(inputValObj)
}

