% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.InterfaceTypeDefinition}
\alias{gqlr_parse.InterfaceTypeDefinition}
\title{parse graphql Definition InterfaceTypeDefinition}
\usage{
\method{gqlr_parse}{InterfaceTypeDefinition}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql Definition InterfaceTypeDefinition
}
\examples{
interfaceObj <- test_json("kitchen_schema")$definitions[[2]]
gqlr_parse(interfaceObj)
}

