% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.ListType}
\alias{gqlr_parse.ListType}
\title{parse graphql Type NonNullType}
\usage{
\method{gqlr_parse}{ListType}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql Type NonNullType
}
\examples{
listObj <- test_json("simple-film-schema")$definitions[[1]]$fields[[3]]$type
gqlr_parse(listObj)
}

