% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.Name}
\alias{gqlr_parse.Name}
\title{parse graphql Name}
\usage{
\method{gqlr_parse}{Name}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql Name
}
\examples{
nameObj <- test_json("simple-film-schema")$definitions[[1]]$name
gqlr_parse(nameObj)
}

