% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.NamedType}
\alias{gqlr_parse.NamedType}
\title{parse graphql Type NamedType}
\usage{
\method{gqlr_parse}{NamedType}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql Type NamedType
}
\examples{
namedTypeObj <- test_json("simple-film-schema")$definitions[[1]]$fields[[1]]$type
gqlr_parse(namedTypeObj)
}

