% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.NonNullType}
\alias{gqlr_parse.NonNullType}
\title{parse graphql Type NonNullType}
\usage{
\method{gqlr_parse}{NonNullType}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql Type NonNullType
}
\examples{
nonNullObj <- test_json("simple-film-schema")$definitions[[1]]$fields[[2]]$type
gqlr_parse(nonNullObj)
}

