% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.ObjectTypeDefinition}
\alias{gqlr_parse.ObjectTypeDefinition}
\title{parse graphql Definition ObjectTypeDefinition}
\usage{
\method{gqlr_parse}{ObjectTypeDefinition}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql Definition ObjectTypeDefinition
}
\examples{
objObj <- test_json("kitchen_schema")$definitions[[1]]
gqlr_parse(objObj)
}

