% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.OperationDefinition}
\alias{gqlr_parse.OperationDefinition}
\title{parse graphql Definition OperationDefinition}
\usage{
\method{gqlr_parse}{OperationDefinition}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql Definition OperationDefinition
}
\examples{
typeExtObj <- test_json("kitchen_schema")$definitions[[7]]
gqlr_parse(typeExtObj)
}

