% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse}
\alias{gqlr_parse}
\title{Generic method to parse graphql json objects}
\usage{
gqlr_parse(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
Generic method to parse graphql json objects
}
\examples{
gqlr_parse(test_json("simple-film-schema"))
}

