% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.ScalarTypeDefinition}
\alias{gqlr_parse.ScalarTypeDefinition}
\title{parse graphql Definition ScalarTypeDefinition}
\usage{
\method{gqlr_parse}{ScalarTypeDefinition}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql Definition ScalarTypeDefinition
}
\examples{
scalarObj <- test_json("kitchen_schema")$definitions[[4]]
gqlr_parse(scalarObj)
}

