% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.SelectionSet}
\alias{gqlr_parse.SelectionSet}
\title{parse graphql SelectionSet}
\usage{
\method{gqlr_parse}{SelectionSet}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql SelectionSet
}
\examples{
typeExtObj <- test_json("kitchen_schema")$definitions[[7]]
gqlr_parse(typeExtObj)
}

