% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.UnionTypeDefinition}
\alias{gqlr_parse.UnionTypeDefinition}
\title{parse graphql Definition UnionTypeDefinition}
\usage{
\method{gqlr_parse}{UnionTypeDefinition}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql Definition UnionTypeDefinition
}
\examples{
unionObj <- test_json("kitchen_schema")$definitions[[3]]
gqlr_parse(unionObj)
}

