% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pryr.R
\name{modify_lang}
\alias{modify_lang}
\title{Recursively modify a language object}
\usage{
modify_lang(x, f, ...)
}
\arguments{
\item{x}{object to modify: should be a call, expression, function or
list of the above.}

\item{f}{function to apply to leaves}

\item{...}{other arguments passed to \code{f}}
}
\value{
Modified language object
}
\description{
Copied from pryr package. Recursively modifies a language object by applying
a function to all leaves.
}
\keyword{internal}
