% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-execution-document.R
\name{parse_selection_set}
\alias{parse_selection_set}
\title{Parse a Selection Set}
\usage{
parse_selection_set(selectionSetObj, fragmentList, contextObj)
}
\description{
Parse a Selection Set
}
\seealso{
https://github.com/facebook/graphql/blob/master/spec/Section%206%20--%20Execution.md#evaluating-selection-sets
}

