% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-scalars.R
\name{scalar_input_boolean}
\alias{scalar_input_boolean}
\alias{scalar_result_boolean}
\title{Scalar Boolean Coercion}
\usage{
scalar_input_boolean(x)

scalar_result_boolean(x)
}
\arguments{
\item{x}{variable in question}
}
\description{
The Boolean scalar type represents true or false. Response formats should use a built-in boolean type if supported; otherwise, they should use their representation of the integers 1 and 0.
}
\details{
Result Coercion

GraphQL servers should coerce non-boolean raw values to Boolean when possible otherwise they must raise a field error. Examples of this may include returning true for any non-zero number.

Input Coercion

When expected as an input type, only boolean input values are accepted. All other input values must raise a query error indicating an incorrect type.
}

