% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-scalars.R
\name{scalar_input_int}
\alias{scalar_input_int}
\alias{scalar_result_int}
\title{Scalar Int Coercion
The Int scalar type represents a signed 32-bit numeric non-fractional values. Response formats that support a 32-bit integer or a number type should use that type to represent this scalar.
Result Coercion
GraphQL servers should coerce non-int raw values to Int when possible otherwise they must raise a field error. Examples of this may include returning 1 for the floating-point number 1.0, or 2 for the string "2".
Input Coercion
When expected as an input type, only integer input values are accepted. All other input values, including strings with numeric content, must raise a query error indicating an incorrect type. If the integer input value represents a value less than -231 or greater than or equal to 231, a query error should be raised.}
\usage{
scalar_input_int(x)

scalar_result_int(x)
}
\arguments{
\item{x}{variable in question}
}
\description{
Note: Numeric integer values larger than 32-bit should either use String or a custom-defined Scalar type, as not all platforms and transports support encoding integer numbers larger than 32-bit.
}

