% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-scalars.R
\name{scalar_input_string}
\alias{scalar_input_string}
\alias{scalar_result_string}
\title{Scalar String Coercion}
\usage{
scalar_input_string(x)

scalar_result_string(x)
}
\arguments{
\item{x}{varible in question}
}
\description{
The String scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text. All response formats must support string representations, and that representation must be used here.
}
\details{
Result Coercion

GraphQL servers should coerce non-string raw values to String when possible otherwise they must raise a field error. Examples of this may include returning the string "true" for a boolean true value, or the string "1" for the integer 1.

Input Coercion

When expected as an input type, only valid UTF-8 string input values are accepted. All other input values must raise a query error indicating an incorrect type.
}

