% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-enum.R
\name{schema_enum}
\alias{schema_enum}
\title{Enums}
\usage{
schema_enum(values)
}
\description{
GraphQL Enums are a variant on the Scalar type, which represents one of a finite set of possible values.
}
\details{
GraphQL Enums are not references for a numeric value, but are unique values in their own right. They serialize as a string: the name of the represented value.

Result Coercion

GraphQL servers must return one of the defined set of possible values. If a reasonable coercion is not possible they must raise a field error.

Input Coercion

GraphQL has a constant literal to represent enum input values. GraphQL string literals must not be accepted as an enum input and instead raise a query error.

Query variable transport serializations which have a different representation for non-string symbolic values (for example, EDN) should only allow such values as enum input values. Otherwise, for most transport serializations that do not, strings may be interpreted as the enum input value with the same name.
}

