% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pryr.R
\name{substitute_q}
\alias{substitute_q}
\title{A version of substitute that evaluates its first argument}
\usage{
substitute_q(x, env)
}
\arguments{
\item{x}{a quoted call}

\item{env}{an environment, or something that behaves like an environment
(like a list or data frame), or a reference to an environment (like a
positive integer or name, see \code{\link{as.environment}} for more
details)}
}
\value{
Modified call with substitutions applied
}
\description{
Copied from pryr package. This version of substitute is needed because
\code{substitute} does not evaluate its first argument, and it's often
useful to be able to modify a quoted call.
}
\keyword{internal}
