% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pryr.R
\name{unenclose}
\alias{unenclose}
\title{Unenclose a closure}
\usage{
unenclose(f)
}
\arguments{
\item{f}{a closure}
}
\value{
A function with values from the enclosing environment substituted
}
\description{
Copied from pryr package. Unenclose a closure by substituting names for
values found in the enclosing environment.
}
\keyword{internal}
