% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{gqlr_parse.Document}
\alias{gqlr_parse.Document}
\title{parse graphql Document}
\usage{
\method{gqlr_parse}{Document}(obj, ...)
}
\arguments{
\item{obj}{object that has class of \code{obj$kind}}

\item{...}{ignored}
}
\description{
parse graphql Document
}
\examples{
gqlr_parse(test_json("simple-film-schema"))
}

